Setup ksp project structure
