+++ ksp/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    kotlin("multiplatform")
+}
+
+group = "com.example"
+version = "1.0-SNAPSHOT"
+
+kotlin {
+    jvm()
+    sourceSets {
+        val jvmMain by getting {
+            dependencies {
+                implementation("com.google.devtools.ksp:symbol-processing-api:1.9.10-1.0.13")
+            }
+            kotlin.srcDir("src/main/kotlin")
+            resources.srcDir("src/main/resources")
+        }
+    }
+}
