+++ ksp/src/main/kotlin/komem/litmus/LitmusTestProcessor.kt
@@ -1,0 +1,16 @@
+package komem.litmus
+
+import com.google.devtools.ksp.processing.*
+import com.google.devtools.ksp.symbol.KSAnnotated
+
+class LitmusTestProcessorProvider : SymbolProcessorProvider {
+    override fun create(environment: SymbolProcessorEnvironment): SymbolProcessor {
+        return LitmusTestProcessor(environment.codeGenerator, environment.logger)
+    }
+}
+
+class LitmusTestProcessor(val codeGenerator: CodeGenerator, val logger: KSPLogger) : SymbolProcessor {
+    override fun process(resolver: Resolver): List<KSAnnotated> {
+        TODO()
+    }
+}
