+++ litmus/src/nativeInterop/kaffinity_gnu.c
