+++ litmus/src/nativeMain/kotlin/komem.litmus/NativeUtils.kt
