+++ settings.gradle.kts
@@ -1,1 +1,4 @@
 rootProject.name = "litmuskt"
+
+include(":litmus")
+include(":ksp")
